# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: Hotel Management API
  version: 1.0.0
  description: |
    API completa para la gestión integral de un sistema hotelero que incluye:
    - Autenticación y gestión de usuarios
    - Administración de habitaciones
    - Gestión de reservaciones y pagos
    - Administración de clientes y usuarios
  contact:
    name: Desarrollador de la API
    email: alfredo.sobrados.gonzalez@gmail.com
  license:
    name: License MIT
    url: https://mit-license.org/
servers:
  - url: http://localhost:8080/api
    description: Entorno de desarrollo local
...
paths:
...
  /admin/rooms:
    get:
      tags: [Rooms]
      summary: Listar todas las habitaciones
      operationId: getAllRooms
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Lista completa de habitaciones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '404':
          description: No hay habitaciones registradas
        '500':
          description: Error interno del servidor
  /admin/room/{id}:
    get:
      tags: [Rooms]
      summary: Obtener una habitación por ID
      operationId: getRoomById
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único de la habitación
          schema:
            type: integer
            example: 101
      responses:
        '200':
          description: Detalles de la habitación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '404':
          description: La habitación solicitada no existe
        '500':
          description: Error interno del servidor
    put:
      tags: [Rooms]
      summary: Actualizar información de una habitación
      operationId: updateRoom
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador de la habitación a actualizar
          schema:
            type: integer
            example: 101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '200':
          description: La actualización se ha hecho correctamente
        '400':
          description: No se ha podido actualizar la habitación
        '500':
          description: Error interno del servidor
  /admin/room:
    post:
      tags: [Rooms]
      summary: Crear una nueva habitación
      operationId: createRoom
      security:
        - bearerAuth: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        '200':
          description: Habitación creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          description: Error en la creación de la habitación
        '500':
          description: Error interno del servidor
  /admin/room/{idDeleteRoom}:
    delete:
      tags: [Rooms]
      summary: Eliminar una habitación por ID
      operationId: deleteRoom
      security:
        - bearerAuth: [admin]
      parameters:
        - name: idDeleteRoom
          in: path
          required: true
          description: ID de la habitación a eliminar
          schema:
            type: integer
            example: 101
      responses:
        '200':
          description: La habitación se ha eliminado correctamente
        '400':
          description: No se ha podido eliminar la habitación
        '500':
          description: Error interno del servidor
  /public/rooms/type/{type}:
    get:
      tags: [Public Rooms]
      summary: Obtener habitaciones por tipo
      operationId: getRoomsByType
      parameters:
        - name: type
          in: path
          required: true
          description: Tipo de habitación a buscar
          schema:
            $ref: '#/components/schemas/RoomType'
      responses:
        '200':
          description: Lista de habitaciones del tipo solicitado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomDTO'
        '404':
          description: No hay habitaciones del tipo solicitado
        '500':
          description: Error interno del servidor
  /public/rooms/available:
    get:
      tags: [Public Rooms]
      summary: Obtener habitaciones disponibles
      operationId: getAvailableRooms
      responses:
        '200':
          description: Lista de habitaciones disponibles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomDTO'
        '404':
          description: No hay habitaciones disponibles
        '500':
          description: Error interno del servidor
  /admin/rooms/{id}/status/{status}:
    put:
      tags: [Rooms]
      summary: Actualizar el estado de una habitación
      operationId: updateStatus
      security:
        - bearerAuth: [admin]
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la habitación a actualizar
          schema:
            type: integer
            example: 101
        - name: status
          in: path
          required: true
          description: Nuevo estado de la habitación
          schema:
            $ref: '#/components/schemas/RoomStatus'
      responses:
        '200':
          description: La actualización se ha hecho correctamente
        '500':
          description: No se ha podido actualizar la habitación o error interno
  /admin/rooms/maintenance:
    get:
      tags: [Rooms]
      summary: Obtener habitaciones en mantenimiento
      operationId: getRoomsInMaintenance
      security:
        - bearerAuth: [admin]
      responses:
        '200':
          description: Lista de habitaciones en mantenimiento
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '404':
          description: No hay habitaciones en mantenimiento
        '500':
          description: Error interno del servidor

  # ==============================================
  # Reservations Endpoints
  # ==============================================
...
components:
  schemas:
...
    Room:
      type: object
      properties:
        id:
          type: integer
          example: 101
        type:
          type: string
          enum: [SINGLE, DOUBLE, SUITE]
          example: DOUBLE
        status:
          type: string
          enum: [AVAILABLE, OCCUPIED, MAINTENANCE]
          example: AVAILABLE
        price:
          type: number
          format: double
          example: 150.50
      required:
        - id
        - type
        - status
        - price
    RoomDTO:
      type: object
      properties:
        id:
          type: integer
          example: 101
        type:
          type: string
          enum: [SINGLE, DOUBLE, SUITE]
          example: SINGLE
        status:
          type: string
          enum: [AVAILABLE, OCCUPIED, MAINTENANCE]
          example: AVAILABLE
      required:
        - id
        - type
        - status
    RoomType:
      type: string
      enum: [SINGLE, DOUBLE, SUITE]
      example: SUITE
    RoomStatus:
      type: string
      enum: [AVAILABLE, OCCUPIED, MAINTENANCE]
      example: AVAILABLE

    # Payments and Reservations
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []